+++ batcave/data.py
@@ -83,7 +83,7 @@ import xml.etree.ElementTree as xml_etree
 # Import internal modules
 from .lang import DEFAULT_ENCODING, BatCaveError, BatCaveException, PathName
 
-SourceType = Enum('SourceType', ('text', 'ini', 'pickle', 'xml_single', 'xml_flat', 'xml'))
+SourceType = Enum('SourceType', ('text', 'ini', 'pickle', 'xml'))
 
 
 class DataError(BatCaveException):
@@ -198,11 +198,10 @@ class DataSource:
                 self._source = {}
             case SourceType.ini:
                 self._source = RawConfigParser()
-            case SourceType.xml_single | SourceType.xml_flat | SourceType.xml:
+            case SourceType.xml:
                 self._source = xml_etree.Element(self.name)
                 self._connection = xml_etree.ElementTree(self._source)
-        if self.type != SourceType.xml_single:
-            self.add_table(self.INFO_TABLE).add_row(schema=str(self._schema))
+        self.add_table(self.INFO_TABLE).add_row(schema=str(self._schema))
         self.commit()
 
     def _load(self) -> None:  # pylint: disable=too-many-branches
@@ -240,7 +239,7 @@ class DataSource:
                 self._source = RawConfigParser()
                 with open(self._connect_info, encoding=DEFAULT_ENCODING) as ini_tmp:
                     self._source.read_file(ini_tmp)
-            case SourceType.xml_single | SourceType.xml_flat | SourceType.xml:
+            case SourceType.xml:
                 if isinstance(self._connect_info, list):
                     self._connection = xml_etree.ElementTree(xml_etree.fromstring(' '.join(self._connect_info)))
                 else:
@@ -328,10 +327,6 @@ class DataSource:
             case SourceType.ini:
                 self._source.add_section(name)
                 self._source.set(name, self.INI_ROW_LIST_OPT, '')
-            case SourceType.xml_single:
-                raise DataError(DataError.INVALID_OPERATION, function='add_table', source_type=self.type)
-            case SourceType.xml_flat:
-                pass
             case SourceType.xml:
                 table = xml_etree.SubElement(self._source, self._XML_TABLE_TAG)
                 table.attrib[self._XML_TABLE_NAME_ATTRIBUTE] = name
@@ -370,7 +365,7 @@ class DataSource:
                 self._connection = open(self._connect_info, 'w', encoding=DEFAULT_ENCODING)  # pylint: disable=consider-using-with
                 self._source.write(self._connection)
                 self._connection.close()
-            case SourceType.xml_single | SourceType.xml_flat | SourceType.xml:
+            case SourceType.xml:
                 cast(ElementTree, self._connection).write(cast(str, self._connect_info), 'ISO-8859-1')
 
     def get_table(self, name: str, /) -> 'DataTable':  # pylint: disable=too-many-branches
@@ -394,11 +389,6 @@ class DataSource:
                 if self._source.has_section(name):
                     row_list = self._source.get(name, self.INI_ROW_LIST_OPT)
                     table = [int(r) for r in row_list.split(',')] if row_list else []
-            case SourceType.xml_single:
-                if name == self.name:
-                    table = self._source
-            case SourceType.xml_flat:
-                table = self._source.findall(name)
             case SourceType.xml:
                 for tmp_table in self._source.iter(self._XML_TABLE_TAG):
                     if tmp_table.get(self._XML_TABLE_NAME_ATTRIBUTE) == name:
@@ -420,10 +410,6 @@ class DataSource:
                 table_names = [t for t in self._source]  # pylint: disable=unnecessary-comprehension
             case SourceType.ini:
                 table_names = [t for t in self._source.sections() if self.INI_ROW_TAG not in t]
-            case SourceType.xml_single:
-                table_names = [self.name]
-            case SourceType.xml_flat:
-                table_names = self._source.findall()
             case SourceType.xml:
                 table_names = [t.get(self._XML_TABLE_NAME_ATTRIBUTE) for t in self._source.iter(self._XML_TABLE_TAG)]
         return [self.get_table(t) for t in table_names]
@@ -444,11 +430,6 @@ class DataSource:
             case SourceType.ini:
                 if self._source.has_section(name):
                     return True
-            case SourceType.xml_single:
-                if name == self.name:
-                    return True
-            case SourceType.xml_flat:
-                return True
             case SourceType.xml:
                 for tmp_table in self._source.iter(self._XML_TABLE_TAG):
                     if tmp_table.get(self._XML_TABLE_NAME_ATTRIBUTE) == name:
@@ -504,15 +485,8 @@ class DataRow:
                 del self._row[col]
             case SourceType.ini:
                 self._parent.remove_option(col)
-            case SourceType.xml_single:
-                if col == self._XML_SINGLE_COL_NAME:
-                    del self._row.attrib[col]
-                else:
-                    self._row.remove(self._row.find(col))
             case SourceType.xml:
                 self._row.remove(self._row.find(col))
-            case SourceType.xml_flat:
-                del self._row.attrib[col]
 
     def delete(self) -> None:
         """Delete this data row.
@@ -537,13 +511,6 @@ class DataRow:
                 return list(self._row.keys())
             case SourceType.ini:
                 return self._parent.options(self._row)
-            case SourceType.xml_single:
-                cols = [e.tag for e in self._row.getiterator()]
-                if self._row.get(self._XML_SINGLE_COL_NAME):
-                    cols.append(self._XML_SINGLE_COL_NAME)
-                return cols
-            case SourceType.xml_flat:
-                return list(self._row.attrib.keys())
             case SourceType.xml:
                 return [e.tag for e in list(self._row) if e.tag != self._XML_ROW_TAG]
         return []
@@ -564,11 +531,6 @@ class DataRow:
             case SourceType.ini:
                 if self.has_col(col):
                     return self._parent.get(self._row, col)
-            case SourceType.xml_single:
-                if col == self._XML_SINGLE_COL_NAME:
-                    return self._row.get(col)
-            case SourceType.xml_flat:
-                return self._row.get(col)
             case SourceType.xml:
                 return self._row.findtext(col)
         return ''
@@ -587,14 +549,10 @@ class DataRow:
                 return col in self._row
             case SourceType.ini:
                 return self._parent.has_option(self._row, col)
-            case SourceType.xml_single | SourceType.xml:
-                if (self.type == SourceType.xml_single) and (col == self._XML_SINGLE_COL_NAME):
-                    return bool(self._row.get(col))
+            case SourceType.xml:
                 if self._row.find(col) is not None:
                     return True
                 return False
-            case SourceType.xml_flat:
-                return bool(self._row.get(col))
         return False
 
     def setvalue(self, col: str, value: str, /) -> None:
@@ -612,16 +570,11 @@ class DataRow:
                 self._row[col] = value
             case SourceType.ini:
                 self._parent.set(self._row, col, value)
-            case SourceType.xml_single | SourceType.xml:
-                if (self.type == SourceType.xml_single) and (col == self._XML_SINGLE_COL_NAME):
-                    self._row.attrib[col] = value
-                if not ((self.type == SourceType.xml_single) and (col == self._XML_SINGLE_COL_NAME)):
-                    col_ref = self._row.find(col)
-                    if col_ref is None:
-                        col_ref = xml_etree.SubElement(self._row, col)
-                    col_ref.text = value
-            case SourceType.xml_flat:
-                self._row.attrib[col] = value
+            case SourceType.xml:
+                col_ref = self._row.find(col)
+                if col_ref is None:
+                    col_ref = xml_etree.SubElement(self._row, col)
+                col_ref.text = value
 
 
 class DataTable:
@@ -675,7 +628,7 @@ class DataTable:
         match self.type:
             case SourceType.text | SourceType.pickle | SourceType.xml:
                 return self._table
-            case SourceType.ini | SourceType.xml_single | SourceType.xml_flat:
+            case SourceType.ini:
                 return self._parent
         raise DataError(DataError.INVALID_TYPE)
 
@@ -700,10 +653,6 @@ class DataTable:
                 self._table.append(row_num)
                 self._parent.set(self.name, DataSource.INI_ROW_LIST_OPT, ','.join([str(r) for r in self._table]))
                 raw_row = self._INI_ROW_FORMAT % (self.name, row_num)
-            case SourceType.xml_single:
-                raw_row = xml_etree.SubElement(self._parent, self._XML_SINGLE_ROW_TAG)
-            case SourceType.xml_flat:
-                raw_row = xml_etree.SubElement(self._parent, self.name)
             case SourceType.xml:
                 raw_row = xml_etree.SubElement(self._table, self._XML_ROW_TAG)
         row = DataRow(self.type, raw_row, self._get_row_parent())