+++ CHANGELOG.md
@@ -16,7 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Upgrade minimum Python to 3.10. (GitLab #62)
   - Replaced switch with new match syntax. (GitLab #62)
 
-<!-- This comment will prevent merge conflicts -->
+- Removed
+  - Removed switch support. (GitLab #66)
+  - Removed xml_flat and xml_single support from data module. (GitLab #66)
+
 ## Release History
 
 ### [41.0.0] - 2022-09-20